+++ client-side/CHANGELOG.md
@@ -3,4 +3,4 @@
 ## [1.0.0] 2018-06-26
 ### Original Release
 - Added Reactstrap as base framework
-- Added design from Paper Kit 2 by Creative Tim
+- Added design from Paper Kit 2 by Shukri Abdi and Bryan Brinson